mex interface

+++ src/tutorial/helmholtz_bem_3d_matrices_gauss.mex.cpp
@@ -41,8 +41,8 @@ void mexFunction(int nlhs, mxArray *lhs[], int nrhs, mxArray const *rhs[])
 	
 	auto const &field_sp = NiHu::dirac(NiHu::constant_view(field_mesh));
 
-	int n = surf_sp.get_num_dofs();
-	int m = field_sp.get_num_dofs();
+	size_t n = surf_sp.get_num_dofs();
+	size_t m = field_sp.get_num_dofs();
 	cMatrix
 		L_surf(n, n, lhs[0]), M_surf(n, n, lhs[1]),
 		Mt_surf(n, n, lhs[2]), D_surf(n, n, lhs[3]),
